Can connect new (temporary) 50 classes onnx
